+++ src/JK.Common.EntityFrameworkCore.SqlServer/Extensions/ModelConfigurationBuilderExtensions.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace JK.Common.EntityFrameworkCore.SqlServer.Extensions;
+﻿namespace JK.Common.EntityFrameworkCore.SqlServer.Extensions;
 
 public static class ModelConfigurationBuilderExtensions
 {
@@ -8,5 +6,7 @@ public static class ModelConfigurationBuilderExtensions
     {
         configurationBuilder.Properties<DateTime>().HaveColumnType("datetime2").HavePrecision(2);
         configurationBuilder.Properties<DateTimeOffset>().HavePrecision(2);
+        configurationBuilder.Properties<decimal>().HavePrecision(19, 5);
+        configurationBuilder.Properties<string>().AreUnicode();
     }
 }