+++ src/JK.Common/Extensions/ObjectExtensions.cs
@@ -6,18 +6,16 @@
 public static class ObjectExtensions
 {
     /// <summary>
-    /// Determines whether the specified value is null.
+    /// Determines whether the object is null.
     /// </summary>
-    /// <typeparam name="T">The type of the value.</typeparam>
     /// <param name="value">The value to check.</param>
     /// <returns><c>true</c> if the value is null; otherwise, <c>false</c>.</returns>
-    public static bool IsNull<T>(this T value) => value is null;
+    public static bool IsNull(this object value) => value is null;
 
     /// <summary>
-    /// Determines whether the specified value is not null.
+    /// Determines whether the object is not null.
     /// </summary>
-    /// <typeparam name="T">The type of the value.</typeparam>
     /// <param name="value">The value to check.</param>
     /// <returns><c>true</c> if the value is not null; otherwise, <c>false</c>.</returns>
-    public static bool IsNotNull<T>(this T value) => value is not null;
+    public static bool IsNotNull(this object value) => value is not null;
 }