+++ src/JK.Common.EntityFrameworkCore/CHANGELOG.md
@@ -12,6 +12,7 @@ Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.co
 ### Added
 
 - Added support for .NET 10.
+- Added `ModelBuilderExtensions` extension method `ApplyAllAssemblyConfigurations()`
 
 ## 5.0.0